% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paths.R
\name{cygwinify_path}
\alias{cygwinify_path}
\title{Make a path suitable for cygwin}
\usage{
cygwinify_path(x = dir())
}
\arguments{
\item{x}{A character vector of file paths. Defaults to files in the
current directory.}
}
\value{
A character vector of the cygwinified inputs.
}
\description{
By default, cygwin complains about standard paths.  This function converts
paths to a form that cygwin likes.
}
\examples{
\dontrun{
cygwinify_path(c("c:/Program Files", "\\\\\\\\some/network/drive"))
}
}
\seealso{
\code{standardize_path}
}

