% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_data.R
\name{dir_data}
\alias{dir_data}
\title{List directory contents in a tidy data frame}
\usage{
dir_data(x = ".", pattern = NULL, all.files = FALSE)
}
\description{
Wrapper to \code{\link[base]{dir}} that returns the results as a tidy data 
frame, for easier \code{\link[utils]{View}}ing.
}
\examples{
base_r_files <- dir_data(R.home(), pattern = "\\\\.R$")
\donttest{
  # Viewing not needed for testing purposes
  utils::View(base_r_files)
}
}
\seealso{
\code{\link[base]{dir}}, \code{\link{split_path}}
}

