% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_libraries}
\alias{get_libraries}
\title{Get the libraries on your machine}
\usage{
get_libraries(index = TRUE, sep = c("/", "\\\\"))
}
\arguments{
\item{index}{A numeric or logical vector specifying the index of the 
libraries to return.  By default, all libraries are returned.}

\item{sep}{String separator between directory levels in the output.}
}
\value{
A character vector of paths to libraries.
}
\description{
Wrapper to \code{\link[base]{.libPaths}} that gets all the libraries that R 
knows about on your machine.
}
\examples{
get_libraries()
get_libraries(1)
}
\references{
\url{http://cran.r-project.org/doc/FAQ/R-FAQ.html#What-is-the-difference-between-package-and-library_003f}
}
\seealso{
\code{\link[base]{.libPaths}}
}

