% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{get_windows_drive}
\alias{get_drive}
\alias{get_windows_drive}
\title{On Windows, return the drive of the path}
\usage{
get_windows_drive(x = getwd())

get_drive(x = getwd())
}
\arguments{
\item{x}{A character vector of file paths. Defaults to the current directory.}
}
\value{
A character vector of drive paths on Windows systems, or forward 
slashes on Unix-based systems.
}
\description{
On a Windows system, this returns the drive letter of the path followed by a 
colon.  On other systems, it returns a single forward slash.
}
\examples{
\donttest{ # often takes > 5s to run
get_windows_drive(c(".", "~", r_home(), temp_dir(), "\\\\\\\\foo/bar"))
}
}
\seealso{
\code{\link{is_windows_drive}}
}

