% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_files}
\alias{list_files}
\title{List the Files in a Directory/Folder}
\usage{
list_files(path = ".", pattern = NULL, all.files = FALSE,
  full.names = TRUE, recursive = FALSE, ignore.case = FALSE,
  include.dirs = FALSE, no.. = FALSE)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link{getwd}()}.  Tilde
    expansion (see \code{\link{path.expand}}) is performed.  Missing
    values will be ignored.}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{all.files}{a logical value.  If \code{FALSE}, only the
    names of visible files are returned.  If \code{TRUE},
    all file names will be returned.}

\item{full.names}{see `list.files`, but set to `TRUE` by default}

\item{recursive}{logical.  Should the listing recurse into directories?}

\item{ignore.case}{logical.  Should pattern-matching be case-insensitive?}

\item{include.dirs}{logical.  Should subdirectory names be included in
    recursive listings?  (They always are in non-recursive ones).}

\item{no..}{logical.  Should both \code{"."} and \code{".."} be excluded
    also from non-recursive listings?}
}
\value{
data frame (tibble form)
}
\description{
This function extends the `base` function `list.files` to return a data frame of the full path
name of a list of files.
}
\details{
The result is a tibble data frame with a single column named `fullname`.
}
\examples{
list_files()
list_files(system.file(package = "pathological"))
list_files(system.file(package = "base"), full.names = FALSE)
}
