% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os-path.R
\name{os_path}
\alias{os_path}
\title{The OS path}
\usage{
os_path(sep = c("/", "\\\\"), standardize = TRUE, splitter = if
  (is_windows()) ";" else ":")
}
\arguments{
\item{sep}{String separator between directory levels in the output.}

\item{standardize}{Should the paths be standardized?}

\item{splitter}{The character to split the PATH environment variable on.
Defaults to a semi-colon on Windows systems and a colon elsewhere.}
}
\value{
A character vector of paths.
}
\description{
The locations in the operating system \code{PATH} environment variable.
}
\examples{
os_path()
}
\seealso{
\code{\link[base]{Sys.getenv}}
}

