% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paths.R
\name{r_profile}
\alias{r_profile}
\alias{r_profile_site}
\title{Get the location of the R profile}
\usage{
r_profile(sep = c("/", "\\\\"))

r_profile_site(sep = c("/", "\\\\"))
}
\arguments{
\item{sep}{String separator between directory levels in the output.}
}
\value{
A string giving the path the \code{".Rprofile"} or
\code{"Rprofile.site"}.  If the file cannot be found, NA is returned.
}
\description{
Gets the location of the user of site R profile startup file.
}
\examples{
r_profile()
r_profile_site()
r_profile("\\\\")
r_profile_site("\\\\")
}
\seealso{
\code{\link[base]{Startup}} for how this is calculated.
}

