% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{sys_which}
\alias{sys_which}
\title{Find paths to executables}
\usage{
sys_which(x, sep = c("/", "\\\\"))
}
\arguments{
\item{x}{A character vector of executables.}

\item{sep}{String separator between directory levels in the output.}
}
\value{
A character vector of paths to those executables, or \code{NA} if it 
doesn't exist. (This behaviour for missing executables differs from 
\code{Sys.which}.)
}
\description{
Wrapper to \code{Sys.which}, that returns standardized paths.
}
\examples{
sys_which("R")              # R executable
sys_which(c("make", "gcc")) # tools for running Rcpp
}
\seealso{
\code{\link[base]{Sys.which}}
}

