% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{temp_dir}
\alias{temp_dir}
\alias{temp_file}
\title{Create a temp file/dir}
\usage{
temp_dir(sep = c("/", "\\\\"))

temp_file(..., sep = c("/", "\\\\"))
}
\arguments{
\item{sep}{String separator between directory levels in the output.}

\item{...}{Passed to \code{tempfile}}
}
\value{
For \code{temp_file} a character vector giving the names of possible 
(temporary) files. Note that no files are generated by \code{temp_file}.
For \code{temp_dir}, the path of the per-session temporary directory.
}
\description{
Wrappers to \code{tempdir} and \code{tempfile} that return standardized 
paths.
}
\examples{
temp_dir()
temp_file()
}
\seealso{
\code{\link[base]{tempdir}}
}

