% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paths.R
\name{r_environ}
\alias{environ}
\alias{r_environ}
\alias{r_environ_site}
\alias{r_profile}
\alias{r_profile_site}
\alias{startup}
\title{Get the location of the R profile/environ}
\usage{
r_environ(sep = c("/", "\\\\"))

r_environ_site(sep = c("/", "\\\\"))

r_profile(sep = c("/", "\\\\"))

r_profile_site(sep = c("/", "\\\\"))
}
\arguments{
\item{sep}{String separator between directory levels in the output.}
}
\value{
A string giving the path the \code{".Rprofile"}, \code{".Renviron"},
\code{"Rprofile.site"}, or \code{".Renviron.site"}.  If the file cannot be
found, NA is returned.
}
\description{
Gets the location of the user or site R profile and environ startup files.
}
\examples{
r_environ()
r_environ_site()
r_profile()
r_profile_site()
}
\seealso{
\code{\link[base]{Startup}} for how this is calculated.
}

